Sort out serial numbers
GnssReceiverType and GnssAntennaType have a SerialNumber
(defined on it) and a
manufacturerModel (defined on BaseGeodeticEquipmentType,
which is the Grandparent of those).  We are uncertain exactly
how to handle this multiplicity of what seems to be an
element with the same purpose. For the mean time we will
have both elements. This code will make sure of this requirement.
It will make an empty one the same as the other or "" if neither
is defined.